\name{TextIO-package}
\alias{TextIO-package}
\alias{TextIO}
\docType{package}
\title{
Generic helper functions for avant-gardED's TextIO project
}
\description{
Use these functions to help with some basic and repeatable tasks that are projected to occur frequently during the TextIO project.
}
\details{
\tabular{ll}{
Package: \tab TextIO\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-17\cr
License: \tab NA\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Nate Giraldi, avant-gardED

Maintainer: <nate@avant-garded.com>

}
\references{
NA
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
